Split out planSlugToPlanProduct
